updates to stream rather than load

+++ Petes_Multi-Systems_Finder.py
@@ -32,6 +32,70 @@ gzip_url_Spansh = "https://downloads.spansh.co.uk/galaxy_populated.json.gz"
 SysDataFilezip = "systemsPopulated.json.gz"
 SysDataFile = "systemsPopulated.json"
 
+def _stream_json_array(file_obj, chunk_size=1024 * 1024, progress_cb=None):
+    decoder = json.JSONDecoder()
+    buffer = ""
+    in_array = False
+    eof = False
+    total_read = 0
+    need_more = True
+    while True:
+        if not eof and (need_more or len(buffer) < chunk_size):
+            data = file_obj.read(chunk_size)
+            if data == "":
+                eof = True
+            else:
+                buffer += data
+                total_read += len(data)
+                if progress_cb:
+                    progress_cb(total_read)
+            need_more = False
+        if not in_array:
+            idx = buffer.find("[")
+            if idx == -1:
+                if eof:
+                    return
+                continue
+            buffer = buffer[idx + 1:]
+            in_array = True
+        buffer = buffer.lstrip()
+        if buffer.startswith("]"):
+            return
+        if buffer.startswith(","):
+            buffer = buffer[1:]
+            continue
+        try:
+            obj, idx = decoder.raw_decode(buffer)
+        except json.JSONDecodeError:
+            if eof:
+                raise
+            need_more = True
+            continue
+        yield obj
+        buffer = buffer[idx:]
+
+def _iter_systems(path):
+    # Use utf-8-sig to tolerate BOMs if present
+    file_size = os.path.getsize(path)
+    last_pct = -1
+    last_ts = time.time()
+    def progress(bytes_read):
+        nonlocal last_pct, last_ts
+        if file_size <= 0:
+            return
+        pct = int(bytes_read * 100 / file_size)
+        now = time.time()
+        if pct != last_pct and (pct == 100 or now - last_ts >= 0.5):
+            print("\rLoading JSON: {:3d}%".format(pct), end="", flush=True)
+            last_pct = pct
+            last_ts = now
+    with open(path, "r", encoding="utf-8-sig") as f:
+        for obj in _stream_json_array(f, progress_cb=progress):
+            yield obj
+    if file_size > 0:
+        print("\rLoading JSON: 100%")
+        print()
+
 def MultiSys():
     #Checks to see if saved file is older than last dump    
     DumpSource = input("Press 1 for EDSM or Enter for Spansh\n")
@@ -62,14 +126,6 @@ def MultiSys():
         GetFile(gzip_url)    
     clear = lambda: os.system('cls')
     
-    #quit_program = input("Press q to quit or Enter to search again - ")
-    print('Loading JSON file.')
-    file=open(SysDataFile,'r')    
-    json_data = open(SysDataFile, "r")
-    jsdatar = json_data.read()
-    SysData = json.loads(jsdatar)
-    print('JSON file loaded.')
-    print()
     while True:
         
         print("Multi-System Finder - By: CMDR Longman.P.J")
@@ -82,8 +138,30 @@ def MultiSys():
         iFact = input("Enter the number for your selected search or q to quit:\n")
         if iFact == "q":
             break
+        print('Streaming JSON file.')
         answer = []
-        for system in SysData:
+        system_iter = _iter_systems(SysDataFile)
+        use_tqdm = False
+        try:
+            from tqdm import tqdm
+            system_iter = tqdm(system_iter, desc="Processing systems", unit="sys", mininterval=0.5)
+            use_tqdm = True
+        except ImportError:
+            processed = 0
+            last_ts = time.time()
+            def _progress_iter(it):
+                nonlocal processed, last_ts
+                for item in it:
+                    processed += 1
+                    now = time.time()
+                    if now - last_ts >= 1.0:
+                        print("\rProcessing systems: {}".format(processed), end="", flush=True)
+                        last_ts = now
+                    yield item
+                print("\rProcessing systems: {} (done)".format(processed))
+                print()
+            system_iter = _progress_iter(system_iter)
+        for system in system_iter:
             countFaction = 0
             countSearchFactions = 0
             countStations = 0
@@ -139,6 +217,8 @@ def MultiSys():
                     print (strFacts)
                 print()
                 print()
+        if use_tqdm:
+            system_iter.close()
                     
         if iFact == '1':
             print ("Federation Faction Systems")
@@ -168,35 +248,45 @@ def MultiSys():
             break        
 
 def GetFile(gzipUrl):
-    statusBar = input("Press y to view the download status.\n")
-    if statusBar == 'y':        
-        from tqdm import tqdm
-        print('Downloading latest zip file.') 
-        # URL of the file to be downloaded is defined as gzip_url
-        # send a HTTP request to the server and save the HTTP response in a response object called response
-        response  = requests.get(gzipUrl, stream=True, allow_redirects=True)
-        total_size = int(response.headers.get('content-length', 0))
-        with open(SysDataFilezip, 'wb') as file, tqdm(
-            desc=SysDataFilezip,
-            total=total_size,
-            unit='B',
-            unit_scale=True,
-            unit_divisor=1024,
-        ) as bar:
-            for data in response.iter_content(chunk_size=1024):
-                size = file.write(data)
-                bar.update(size)
+    downloaded = False
+    use_existing = False
+    if os.path.isfile(SysDataFilezip):
+        use_existing_zip = input("Existing zip file found. Press y to use it and skip download, or Enter to download latest.\n")
+        if use_existing_zip.lower() == 'y':
+            use_existing = True
+
+    if use_existing:
+        print('Using existing zip file. Skipping download.')
     else:
-        print('Downloading latest zip file.') 
-        # URL of the file to be downloaded is defined as gzip_url
-        res  = requests.get(gzipUrl) # create HTTP response object
-        # send a HTTP request to the server and save the HTTP response in a response object called r
-        with open(SysDataFilezip,'wb') as f:
-            # Saving received content as a file in binary format  
-            # write the contents of the response (r.content) to a new file in binary mode.
-            f.write(res.content)        
- 
-    print('Zip file downloaded.')
+        statusBar = input("Press y to view the download status.\n")
+        if statusBar == 'y':        
+            from tqdm import tqdm
+            print('Downloading latest zip file.') 
+            # URL of the file to be downloaded is defined as gzip_url
+            # send a HTTP request to the server and save the HTTP response in a response object called response
+            response  = requests.get(gzipUrl, stream=True, allow_redirects=True)
+            total_size = int(response.headers.get('content-length', 0))
+            with open(SysDataFilezip, 'wb') as file, tqdm(
+                desc=SysDataFilezip,
+                total=total_size,
+                unit='B',
+                unit_scale=True,
+                unit_divisor=1024,
+            ) as bar:
+                for data in response.iter_content(chunk_size=1024):
+                    size = file.write(data)
+                    bar.update(size)
+        else:
+            print('Downloading latest zip file.') 
+            # URL of the file to be downloaded is defined as gzip_url
+            res  = requests.get(gzipUrl) # create HTTP response object
+            # send a HTTP request to the server and save the HTTP response in a response object called r
+            with open(SysDataFilezip,'wb') as f:
+                # Saving received content as a file in binary format  
+                # write the contents of the response (r.content) to a new file in binary mode.
+                f.write(res.content)
+        downloaded = True
+        print('Zip file downloaded.')
     print('Extracting zip file.')
 
     #path_to_file_to_be_extracted
@@ -205,16 +295,17 @@ def GetFile(gzipUrl):
 
     #output file to be filled
 
-    op = open(SysDataFile,"w") 
-
-    with gzip.open(ip,"rb") as ip_byte:
-        op.write(ip_byte.read().decode("utf-8"))
-        ip_byte.close()
+    # Stream-decompress to avoid loading the whole file into memory
+    with gzip.open(ip, "rb") as ip_byte, open(SysDataFile, "wb") as op:
+        shutil.copyfileobj(ip_byte, op, length=1024 * 1024)
     print('Saving extracted zip file as JSON file.')
     dir_path = os.path.dirname(os.path.realpath(SysDataFilezip))
-    zippath = dir_path + '\\' + SysDataFilezip
-    os.remove(zippath) #delete zipped file
-    print('Zip file deleted.')
+    zippath = os.path.join(dir_path, SysDataFilezip)
+    if downloaded:
+        os.remove(zippath) #delete zipped file
+        print('Zip file deleted.')
+    else:
+        print('Existing zip file kept.')
 
 if __name__ == "__main__":
     MultiSys()
